+++ src/pug.config.js
@@ -1,5 +1,0 @@
-const locals = require('./tools/meta');
-
-module.exports = {
-  locals,
-};